LayoutTests/platform/chromium/fast/dom/Geolocation/notimer-after-unload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: line 1: Blocked alert('unload-called') during unload.
You should have seen one unload alert appear.
